extract and restructure retries module
