+++ docs/source/retry-policy/downgrading-consistency.md
@@ -54,7 +54,7 @@ To use in `Session`:
 # async fn check_only_compiles() -> Result<(), Box<dyn Error>> {
 use scylla::{Session, SessionBuilder};
 use scylla::execution::ExecutionProfile;
-use scylla::execution::downgrading_consistency_retry_policy::DowngradingConsistencyRetryPolicy;
+use scylla::execution::retries::DowngradingConsistencyRetryPolicy;
 
 let handle = ExecutionProfile::builder()
     .retry_policy(Arc::new(DowngradingConsistencyRetryPolicy::new()))
@@ -79,7 +79,7 @@ To use in a [simple query](../queries/simple.md):
 # async fn check_only_compiles(session: &Session) -> Result<(), Box<dyn Error>> {
 use scylla::query::Query;
 use scylla::execution::ExecutionProfile;
-use scylla::execution::downgrading_consistency_retry_policy::DowngradingConsistencyRetryPolicy;
+use scylla::execution::retries::DowngradingConsistencyRetryPolicy;
 
 let handle = ExecutionProfile::builder()
     .retry_policy(Arc::new(DowngradingConsistencyRetryPolicy::new()))
@@ -106,7 +106,7 @@ To use in a [prepared query](../queries/prepared.md):
 # async fn check_only_compiles(session: &Session) -> Result<(), Box<dyn Error>> {
 use scylla::prepared_statement::PreparedStatement;
 use scylla::execution::ExecutionProfile;
-use scylla::execution::downgrading_consistency_retry_policy::DowngradingConsistencyRetryPolicy;
+use scylla::execution::retries::DowngradingConsistencyRetryPolicy;
 
 let handle = ExecutionProfile::builder()
     .retry_policy(Arc::new(DowngradingConsistencyRetryPolicy::new()))
