+++ scylla/src/execution/retries/default_rp.rs
@@ -1,76 +1,8 @@
-//! Query retries configurations\
-//! To decide when to retry a query the `Session` can use any object which implements
-//! the `RetryPolicy` trait
+use scylla_cql::frame::response::error::{DbError, WriteType};
 
-use crate::frame::types::Consistency;
-use crate::transport::errors::{DbError, QueryError, WriteType};
+use crate::transport::errors::QueryError;
 
-/// Information about a failed query
-pub struct QueryInfo<'a> {
-    /// The error with which the query failed
-    pub error: &'a QueryError,
-    /// A query is idempotent if it can be applied multiple times without changing the result of the initial application\
-    /// If set to `true` we can be sure that it is idempotent\
-    /// If set to `false` it is unknown whether it is idempotent
-    pub is_idempotent: bool,
-    /// Consistency with which the query failed
-    pub consistency: Consistency,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum RetryDecision {
-    RetrySameNode(Option<Consistency>), // None means that the same consistency should be used as before
-    RetryNextNode(Option<Consistency>), // ditto
-    DontRetry,
-    IgnoreWriteError,
-}
-
-/// Specifies a policy used to decide when to retry a query
-pub trait RetryPolicy: std::fmt::Debug + Send + Sync {
-    /// Called for each new query, starts a session of deciding about retries
-    fn new_session(&self) -> Box<dyn RetrySession>;
-}
-
-/// Used throughout a single query to decide when to retry it
-/// After this query is finished it is destroyed or reset
-pub trait RetrySession: Send + Sync {
-    /// Called after the query failed - decide what to do next
-    fn decide_should_retry(&mut self, query_info: QueryInfo) -> RetryDecision;
-
-    /// Reset before using for a new query
-    fn reset(&mut self);
-}
-
-/// Forwards all errors directly to the user, never retries
-#[derive(Debug)]
-pub struct FallthroughRetryPolicy;
-pub struct FallthroughRetrySession;
-
-impl FallthroughRetryPolicy {
-    pub fn new() -> FallthroughRetryPolicy {
-        FallthroughRetryPolicy
-    }
-}
-
-impl Default for FallthroughRetryPolicy {
-    fn default() -> FallthroughRetryPolicy {
-        FallthroughRetryPolicy
-    }
-}
-
-impl RetryPolicy for FallthroughRetryPolicy {
-    fn new_session(&self) -> Box<dyn RetrySession> {
-        Box::new(FallthroughRetrySession)
-    }
-}
-
-impl RetrySession for FallthroughRetrySession {
-    fn decide_should_retry(&mut self, _query_info: QueryInfo) -> RetryDecision {
-        RetryDecision::DontRetry
-    }
-
-    fn reset(&mut self) {}
-}
+use super::{QueryInfo, RetryDecision, RetryPolicy, RetrySession};
 
 /// Default retry policy - retries when there is a high chance that a retry might help.\
 /// Behaviour based on [DataStax Java Driver](https://docs.datastax.com/en/developer/java-driver/4.10/manual/core/retries/)
