+++ scylla/src/statement/mod.rs
@@ -10,7 +10,7 @@ use std::{sync::Arc, time::Duration};
 use thiserror::Error;
 
 use crate::execution::execution_profile::ExecutionProfileHandle;
-use crate::{history::HistoryListener, retry_policy::RetryPolicy};
+use crate::{history::HistoryListener, retries::RetryPolicy};
 
 pub mod batch;
 pub mod prepared_statement;
