+++ scylla/src/statement/prepared_statement.rs
@@ -14,10 +14,10 @@ use uuid::Uuid;
 
 use super::{PageSize, StatementConfig};
 use crate::execution::execution_profile::ExecutionProfileHandle;
+use crate::execution::retries::RetryPolicy;
 use crate::frame::response::result::PreparedMetadata;
 use crate::frame::types::{Consistency, SerialConsistency};
 use crate::history::HistoryListener;
-use crate::retry_policy::RetryPolicy;
 use crate::routing::Token;
 use crate::transport::errors::{BadQuery, ProtocolError, QueryError};
 use crate::transport::partitioner::{Partitioner, PartitionerHasher, PartitionerName};
