+++ scylla/src/statement/query.rs
@@ -1,8 +1,8 @@
 use super::{PageSize, StatementConfig};
 use crate::execution::execution_profile::ExecutionProfileHandle;
+use crate::execution::retries::RetryPolicy;
 use crate::frame::types::{Consistency, SerialConsistency};
 use crate::history::HistoryListener;
-use crate::retry_policy::RetryPolicy;
 use std::sync::Arc;
 use std::time::Duration;
