+++ scylla/tests/integration/retries.rs
@@ -1,6 +1,6 @@
 use crate::utils::{setup_tracing, test_with_3_node_cluster, unique_keyspace_name, PerformDDL};
+use scylla::execution::retries::FallthroughRetryPolicy;
 use scylla::query::Query;
-use scylla::retry_policy::FallthroughRetryPolicy;
 use scylla::session::Session;
 use scylla::speculative_execution::SimpleSpeculativeExecutionPolicy;
 use scylla::ExecutionProfile;
